+++ metadata.rb
@@ -4,11 +4,11 @@ maintainer_email 'ketanpadegaonkar@gmail.com'
 license          'MIT'
 description      'Installs/Configures gauge'
 long_description  IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '0.4.0'
+version          '0.5.0'
 
 
 %w{redhat centos scientific fedora amazon oracle debian ubuntu suse opensuse}.each do |os|
   supports os
 end
 
-suggests 'chocolatey'
+depends 'chocolatey'